Make more extractions
