+++ src/pages/claims/[claimId]/members/[memberId].tsx
@@ -2,19 +2,19 @@ import styled from '@emotion/styled'
 import { Card, CardsWrapper, FadeIn, MainHeadline } from '@hedvig-ui'
 import { ClaimState, useClaimPageQuery } from 'api/generated/graphql'
 import { ClaimItems } from 'components/claims/claim-details/components/claim-items'
+import { ClaimEvents } from 'components/claims/claim-details/components/ClaimEvents'
+import { ClaimFileTable } from 'components/claims/claim-details/components/ClaimFileTable'
+import { ClaimInformation } from 'components/claims/claim-details/components/ClaimInformation'
+import { ClaimNotes } from 'components/claims/claim-details/components/ClaimNotes'
+import { ClaimPayments } from 'components/claims/claim-details/components/ClaimPayments'
+import { ClaimTranscriptions } from 'components/claims/claim-details/components/ClaimTranscriptions'
+import { ClaimTypeForm } from 'components/claims/claim-details/components/ClaimType'
+import { MemberInformation } from 'components/claims/claim-details/components/MemberInformation'
 import { ChatPane } from 'components/member/tabs/ChatPane'
 import React, { useContext, useEffect } from 'react'
 import { Prompt, RouteComponentProps } from 'react-router'
 import { MemberHistoryContext } from 'utils/member-history'
 import { getCarrierText } from 'utils/text'
-import { ClaimEvents } from './components/ClaimEvents'
-import { ClaimFileTable } from './components/ClaimFileTable'
-import { ClaimInformation } from './components/ClaimInformation'
-import { ClaimNotes } from './components/ClaimNotes'
-import { ClaimPayments } from './components/ClaimPayments'
-import { ClaimTranscriptions } from './components/ClaimTranscriptions'
-import { ClaimTypeForm } from './components/ClaimType'
-import { MemberInformation } from './components/MemberInformation'
 
 const ChatPaneAdjustedContainer = styled.div`
   width: clamp(1000px, calc(100% - 400px), calc(100% - 400px));
