+++ src/pages/members/[memberId]/[tab].tsx
@@ -1,8 +1,12 @@
 import { MemberTabs } from 'components/member'
 import { useGetMemberInfo } from 'graphql/use-get-member-info'
 import React from 'react'
+import { RouteComponentProps } from 'react-router'
 
-export const MemberPage = (props) => {
+export const MemberPage: React.FC<RouteComponentProps<{
+  memberId: string
+  tab: string
+}>> = (props) => {
   const memberId = props.match.params.memberId
   const [member] = useGetMemberInfo(memberId)
